Resolved Pipeline failure issue

+++ Src/features/build_features.py
@@ -46,7 +46,7 @@ def build_features(df):
     df["month"] = df["timestamp"].dt.month
     df["weekday"] = df["timestamp"].dt.weekday
     df["dayofyear"] = df["timestamp"].dt.dayofyear
-    df["is_weekend"] = (df["weekday"].isin([5, 6])).astype(int)
+    df["is_weekend"] = (df["weekday"].isin([5, 6])).astype("int32")
 
     # ==================== POLLUTANTS ====================
     pollutants = ["pm2_5", "pm10", "carbon_monoxide", "nitrogen_dioxide", "sulphur_dioxide", "ozone"]
